Simplify dependencies
Tool failed when pandas was not included. Not sure of the current error as I have checked the test files and they have only 26 populations
error: Exit code 1 (There are too many populations in the input files. The maximum number of populations is 40.)', 'exit_code': 1, 'code_desc': 'There are too many populations in the input files. The maximum number of populations is 40.', 'error_level': 3}