Detect alternative spill over matrix name
This commit should detect an alternative name for the spillover matrix stored in the keywords. There has been one set of FCS files I found externally where the spillover matrix was labelled as such. I thought it would be useful to have this PR for in the future, if you come across this issue again. Alternative ways of labelling the spillover matrix could easily be added here as well.